requirements.txt
Django==4.2.7
paho-mqtt==1.6.1
psycopg[binary]
python-dotenv
channels-redis
sentry_sdk #error logging
pytz #timezone
djangorestframework
djangorestframework-simplejwt
coreapi
drf-spectacular==0.28.0
tzdata